--- docs/api/ptk/completer.rst
@@ -1,7 +1,0 @@
-.. _xonsh_ptk_completer:
-
-*************************************************************
-Prompt Toolkit Completer (``xonsh.ptk.completer``)
-*************************************************************
-
-.. warning:: This module is pending deprecation.
