--- docs/api/ptk/history.rst
@@ -1,7 +1,0 @@
-.. _xonsh_ptk_history:
-
-****************************************************************
-Prompt Toolkit History Object (``xonsh.ptk.history``)
-****************************************************************
-
-.. warning:: This module is pending deprecation.
