--- docs/api/ptk/key_bindings.rst
@@ -1,8 +1,0 @@
-.. _xonsh_ptk_key_bindings:
-
-**********************************************************
-Prompt Toolkit Key Bindings (``xonsh.ptk.key_bindings``)
-**********************************************************
-
-.. warning:: This module is pending deprecation.
-
