--- docs/api/ptk/shell.rst
@@ -1,7 +1,0 @@
-.. _xonsh_ptk_shell:
-
-******************************************************
-Prompt Toolkit Shell (``xonsh.ptk.shell``)
-******************************************************
-
-.. warning:: This module is pending deprecation.
