+++ docs/api/ptk_shell/completer.rst
@@ -1,10 +1,10 @@
 .. _xonsh_ptk2_completer:
 
 *************************************************************
-Prompt Toolkit 2 Completer (``xonsh.ptk2.completer``)
+Prompt Toolkit 2+ Completer (``xonsh.ptk_shell.completer``)
 *************************************************************
 
-.. automodule:: xonsh.ptk2.completer
+.. automodule:: xonsh.ptk_shell.completer
     :members:
     :undoc-members:
     :inherited-members:
