+++ docs/api/ptk_shell/key_bindings.rst
@@ -1,0 +1,10 @@
+.. _xonsh_ptk_shell_key_bindings:
+
+*****************************************************************
+Prompt Toolkit 2+ Key Bindings (``xonsh.ptk_shell.key_bindings``)
+*****************************************************************
+
+.. automodule:: xonsh.ptk_shell.key_bindings
+    :members:
+    :undoc-members:
+    :inherited-members:
