+++ docs/conf.py
@@ -32,12 +32,7 @@ spec = importlib.util.find_spec("prompt_toolkit")
 if spec is not None:
     # hacky runaround to import PTK-specific events
     builtins.__xonsh__.env = Env()
-    from xonsh.platform import ptk_version_info
-
-    if ptk_version_info()[0] < 2:
-        from xonsh.ptk.shell import events
-    else:
-        from xonsh.ptk2.shell import events
+    from xonsh.ptk_shell.shell import events
 else:
     from xonsh.events import events
