+++ requirements/tests.txt
@@ -4,7 +4,7 @@ flake8
 pytest-flake8
 pytest-cov
 pytest-timeout
-prompt-toolkit
+prompt-toolkit>=2.0
 pygments>=2.2
 codecov
 coverage
