+++ xonsh/platform.py
@@ -169,19 +169,13 @@ def ptk_version_info():
         return None
 
 
+minimum_required_ptk_version = (2, 0)
+"""Minimum version of prompt-toolkit supported by Xonsh"""
+
+
 @functools.lru_cache(1)
 def ptk_above_min_supported():
-    minimum_required_ptk_version = (1, 0)
-    return ptk_version_info()[:2] >= minimum_required_ptk_version
-
-
-@functools.lru_cache(1)
-def ptk_shell_type():
-    """Returns the prompt_toolkit shell type based on the installed version."""
-    if ptk_version_info()[:2] < (2, 0):
-        return "prompt_toolkit1"
-    else:
-        return "prompt_toolkit2"
+    return ptk_version_info() and ptk_version_info()[:2] >= minimum_required_ptk_version
 
 
 @functools.lru_cache(1)
@@ -198,7 +192,7 @@ def win_ansi_support():
 
 @functools.lru_cache(1)
 def ptk_below_max_supported():
-    ptk_max_version_cutoff = (2, 0)
+    ptk_max_version_cutoff = (99999, 0)  # currently, no limit.
     return ptk_version_info()[:2] < ptk_max_version_cutoff
