+++ xonsh/ptk2/__init__.py
@@ -1,0 +1,1 @@
+from xonsh.ptk_shell import *  # noqa: F403 F401
