+++ xonsh/ptk_shell/completer.py
@@ -3,9 +3,9 @@
 import os
 import builtins
 
-from prompt_toolkit.layout.dimension import LayoutDimension
 from prompt_toolkit.completion import Completer, Completion
 from prompt_toolkit.auto_suggest import AutoSuggestFromHistory
+from prompt_toolkit.application.current import get_app
 
 
 class PromptToolkitCompleter(Completer):
@@ -80,8 +80,8 @@ class PromptToolkitCompleter(Completer):
 
     def suggestion_completion(self, document, line):
         """Provides a completion based on the current auto-suggestion."""
-        cli = self.shell.prompter.cli
-        sug = self.hist_suggester.get_suggestion(cli, cli.current_buffer, document)
+        app = self.shell.prompter.app
+        sug = self.hist_suggester.get_suggestion(app.current_buffer, document)
         if sug is None:
             return None
         comp, _, _ = sug.text.partition(" ")
@@ -89,20 +89,17 @@ class PromptToolkitCompleter(Completer):
         return prev + comp
 
     def reserve_space(self):
-        cli = builtins.__xonsh__.shell.shell.prompter.cli
-        window = cli.application.layout.children[0].content.children[1]
+        """Adjust the height for showing autocompletion menu."""
+        app = get_app()
+        render = app.renderer
+        window = app.layout.container.children[0].content.children[1].content
 
         if window and window.render_info:
             h = window.render_info.content_height
             r = builtins.__xonsh__.env.get("COMPLETIONS_MENU_ROWS")
             size = h + r
-
-            def comp_height(cli):
-                # If there is an autocompletion menu to be shown, make sure that o
-                # layout has at least a minimal height in order to display it.
-                if not cli.is_done:
-                    return LayoutDimension(min=size)
-                else:
-                    return LayoutDimension()
-
-            window._height = comp_height
+            last_h = render._last_screen.height if render._last_screen else 0
+            last_h = max(render._min_available_height, last_h)
+            if last_h < size:
+                if render._last_screen:
+                    render._last_screen.height = size
