+++ xonsh/ptk_shell/shell.py
@@ -1,31 +1,36 @@
 # -*- coding: utf-8 -*-
 """The prompt_toolkit based xonsh shell."""
+import os
 import sys
 import builtins
+from types import MethodType
 
-from prompt_toolkit.key_binding.manager import KeyBindingManager
-from prompt_toolkit.auto_suggest import AutoSuggestFromHistory
-from prompt_toolkit.layout.lexers import PygmentsLexer
-from prompt_toolkit.shortcuts import print_tokens
-from prompt_toolkit.styles import PygmentsStyle, style_from_dict
-
-from xonsh.base_shell import BaseShell
-from xonsh.tools import print_exception, carriage_return, ansicolors_to_ptk1_names
-from xonsh.ptk.completer import PromptToolkitCompleter
-from xonsh.ptk.history import PromptToolkitHistory
-from xonsh.ptk.key_bindings import load_xonsh_bindings
-from xonsh.ptk.shortcuts import Prompter
 from xonsh.events import events
+from xonsh.base_shell import BaseShell
 from xonsh.shell import transform_command
-from xonsh.platform import HAS_PYGMENTS, ON_WINDOWS
-from xonsh.style_tools import (
-    partial_color_tokenize,
-    _TokenType,
-    DEFAULT_STYLE_DICT as _DEFAULT_STYLE_DICT,
-)
+from xonsh.tools import print_exception, carriage_return
+from xonsh.platform import HAS_PYGMENTS, ON_WINDOWS, ON_POSIX
+from xonsh.style_tools import partial_color_tokenize, _TokenType, DEFAULT_STYLE_DICT
 from xonsh.lazyimps import pygments, pyghooks, winutils
 from xonsh.pygments_cache import get_all_styles
-from xonsh.lazyasd import LazyObject
+from xonsh.ptk_shell.history import PromptToolkitHistory, _cust_history_matches
+from xonsh.ptk_shell.completer import PromptToolkitCompleter
+from xonsh.ptk_shell.key_bindings import load_xonsh_bindings
+
+from prompt_toolkit.auto_suggest import AutoSuggestFromHistory
+from prompt_toolkit.lexers import PygmentsLexer
+from prompt_toolkit.enums import EditingMode
+from prompt_toolkit.key_binding import KeyBindings
+from prompt_toolkit.history import ThreadedHistory
+from prompt_toolkit.shortcuts import print_formatted_text as ptk_print
+from prompt_toolkit.shortcuts import CompleteStyle
+from prompt_toolkit.shortcuts.prompt import PromptSession
+from prompt_toolkit.formatted_text import PygmentsTokens
+from prompt_toolkit.styles import merge_styles, Style
+from prompt_toolkit.styles.pygments import (
+    style_from_pygments_cls,
+    style_from_pygments_dict,
+)
 
 
 Token = _TokenType()
@@ -34,54 +39,45 @@ events.transmogrify("on_ptk_create", "LoadEvent")
 events.doc(
     "on_ptk_create",
     """
-on_ptk_create(prompter: Prompter, history: PromptToolkitHistory, completer: PromptToolkitCompleter, bindings: KeyBindingManager) ->
+on_ptk_create(prompter: PromptSession, history: PromptToolkitHistory, completer: PromptToolkitCompleter, bindings: KeyBindings) ->
 
 Fired after prompt toolkit has been initialized
 """,
 )
 
-# Convert new ansicolor names to names
-# understood by PTK1
-DEFAULT_STYLE_DICT = LazyObject(
-    lambda: ansicolors_to_ptk1_names(_DEFAULT_STYLE_DICT),
-    globals(),
-    "DEFAULT_STYLE_DICT",
-)
-
 
 class PromptToolkitShell(BaseShell):
-    """The xonsh shell."""
+    """The xonsh shell for prompt_toolkit v2 and later."""
+
+    completion_displays_to_styles = {
+        "multi": CompleteStyle.MULTI_COLUMN,
+        "single": CompleteStyle.COLUMN,
+        "readline": CompleteStyle.READLINE_LIKE,
+        "none": None,
+    }
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         if ON_WINDOWS:
             winutils.enable_virtual_terminal_processing()
         self._first_prompt = True
-        self.prompter = Prompter()
-        self.history = PromptToolkitHistory()
+        self.history = ThreadedHistory(PromptToolkitHistory())
+        self.prompter = PromptSession(history=self.history)
         self.pt_completer = PromptToolkitCompleter(self.completer, self.ctx, self)
-        key_bindings_manager_args = {
-            "enable_auto_suggest_bindings": True,
-            "enable_search": True,
-            "enable_abort_and_exit_bindings": True,
-        }
-        self.key_bindings_manager = KeyBindingManager(**key_bindings_manager_args)
-        load_xonsh_bindings(self.key_bindings_manager)
+        self.key_bindings = KeyBindings()
+        load_xonsh_bindings(self.key_bindings)
+        # Store original `_history_matches` in case we need to restore it
+        self._history_matches_orig = self.prompter.default_buffer._history_matches
         # This assumes that PromptToolkitShell is a singleton
         events.on_ptk_create.fire(
             prompter=self.prompter,
             history=self.history,
             completer=self.pt_completer,
-            bindings=self.key_bindings_manager,
+            bindings=self.key_bindings,
         )
 
     def singleline(
-        self,
-        store_in_history=True,
-        auto_suggest=None,
-        enable_history_search=True,
-        multiline=True,
-        **kwargs
+        self, auto_suggest=None, enable_history_search=True, multiline=True, **kwargs
     ):
         """Reads a single line of input from the shell. The store_in_history
         kwarg flags whether the input should be stored in PTK's in-memory
@@ -90,14 +86,13 @@ class PromptToolkitShell(BaseShell):
         events.on_pre_prompt.fire()
         env = builtins.__xonsh__.env
         mouse_support = env.get("MOUSE_SUPPORT")
-        if store_in_history:
-            history = self.history
-        else:
-            history = None
-            enable_history_search = False
         auto_suggest = auto_suggest if env.get("AUTO_SUGGEST") else None
+        refresh_interval = env.get("PROMPT_REFRESH_INTERVAL")
+        refresh_interval = refresh_interval if refresh_interval > 0 else None
+        complete_in_thread = env.get("COMPLETION_IN_THREAD")
         completions_display = env.get("COMPLETIONS_DISPLAY")
-        multicolumn = completions_display == "multi"
+        complete_style = self.completion_displays_to_styles[completions_display]
+
         complete_while_typing = env.get("UPDATE_COMPLETIONS_ON_KEYPRESS")
         if complete_while_typing:
             # PTK requires history search to be none when completing while typing
@@ -105,45 +100,71 @@ class PromptToolkitShell(BaseShell):
         if HAS_PYGMENTS:
             self.styler.style_name = env.get("XONSH_COLOR_STYLE")
         completer = None if completions_display == "none" else self.pt_completer
-        if not env.get("UPDATE_PROMPT_ON_KEYPRESS"):
-            prompt_tokens_cached = self.prompt_tokens(None)
-            get_prompt_tokens = lambda cli: prompt_tokens_cached
-            rprompt_tokens_cached = self.rprompt_tokens(None)
-            get_rprompt_tokens = lambda cli: rprompt_tokens_cached
-            bottom_toolbar_tokens_cached = self.bottom_toolbar_tokens(None)
-            get_bottom_toolbar_tokens = lambda cli: bottom_toolbar_tokens_cached
-        else:
+
+        get_bottom_toolbar_tokens = self.bottom_toolbar_tokens
+
+        if env.get("UPDATE_PROMPT_ON_KEYPRESS"):
             get_prompt_tokens = self.prompt_tokens
             get_rprompt_tokens = self.rprompt_tokens
-            get_bottom_toolbar_tokens = self.bottom_toolbar_tokens
+        else:
+            get_prompt_tokens = self.prompt_tokens()
+            get_rprompt_tokens = self.rprompt_tokens()
+            if get_bottom_toolbar_tokens:
+                get_bottom_toolbar_tokens = get_bottom_toolbar_tokens()
 
-        with self.prompter:
-            prompt_args = {
-                "mouse_support": mouse_support,
-                "auto_suggest": auto_suggest,
-                "get_prompt_tokens": get_prompt_tokens,
-                "get_rprompt_tokens": get_rprompt_tokens,
-                "get_bottom_toolbar_tokens": get_bottom_toolbar_tokens,
-                "completer": completer,
-                "multiline": multiline,
-                "get_continuation_tokens": self.continuation_tokens,
-                "history": history,
-                "enable_history_search": enable_history_search,
-                "reserve_space_for_menu": 0,
-                "key_bindings_registry": self.key_bindings_manager.registry,
-                "display_completions_in_columns": multicolumn,
-                "complete_while_typing": complete_while_typing,
-            }
-            if builtins.__xonsh__.env.get("COLOR_INPUT"):
-                if HAS_PYGMENTS:
-                    prompt_args["lexer"] = PygmentsLexer(pyghooks.XonshLexer)
-                    prompt_args["style"] = PygmentsStyle(
-                        pyghooks.xonsh_style_proxy(self.styler)
-                    )
-                else:
-                    prompt_args["style"] = style_from_dict(DEFAULT_STYLE_DICT)
-            line = self.prompter.prompt(**prompt_args)
-            events.on_post_prompt.fire()
+        if env.get("VI_MODE"):
+            editing_mode = EditingMode.VI
+        else:
+            editing_mode = EditingMode.EMACS
+
+        if env.get("XONSH_HISTORY_MATCH_ANYWHERE"):
+            self.prompter.default_buffer._history_matches = MethodType(
+                _cust_history_matches, self.prompter.default_buffer
+            )
+        elif (
+            self.prompter.default_buffer._history_matches
+            is not self._history_matches_orig
+        ):
+            self.prompter.default_buffer._history_matches = self._history_matches_orig
+
+        prompt_args = {
+            "mouse_support": mouse_support,
+            "auto_suggest": auto_suggest,
+            "message": get_prompt_tokens,
+            "rprompt": get_rprompt_tokens,
+            "bottom_toolbar": get_bottom_toolbar_tokens,
+            "completer": completer,
+            "multiline": multiline,
+            "editing_mode": editing_mode,
+            "prompt_continuation": self.continuation_tokens,
+            "enable_history_search": enable_history_search,
+            "reserve_space_for_menu": 0,
+            "key_bindings": self.key_bindings,
+            "complete_style": complete_style,
+            "complete_while_typing": complete_while_typing,
+            "include_default_pygments_style": False,
+            "refresh_interval": refresh_interval,
+            "complete_in_thread": complete_in_thread,
+        }
+        if builtins.__xonsh__.env.get("COLOR_INPUT"):
+            if HAS_PYGMENTS:
+                prompt_args["lexer"] = PygmentsLexer(pyghooks.XonshLexer)
+                style = style_from_pygments_cls(pyghooks.xonsh_style_proxy(self.styler))
+            else:
+                style = style_from_pygments_dict(DEFAULT_STYLE_DICT)
+
+            prompt_args["style"] = style
+
+            style_overrides_env = env.get("PTK_STYLE_OVERRIDES")
+            if style_overrides_env:
+                try:
+                    style_overrides = Style.from_dict(style_overrides_env)
+                    prompt_args["style"] = merge_styles([style, style_overrides])
+                except (AttributeError, TypeError, ValueError):
+                    print_exception()
+
+        line = self.prompter.prompt(**prompt_args)
+        events.on_post_prompt.fire()
         return line
 
     def _push(self, line):
@@ -187,7 +208,7 @@ class PromptToolkitShell(BaseShell):
                 else:
                     break
 
-    def prompt_tokens(self, cli):
+    def prompt_tokens(self):
         """Returns a list of (token, str) tuples for the current prompt."""
         p = builtins.__xonsh__.env.get("PROMPT")
         try:
@@ -199,9 +220,9 @@ class PromptToolkitShell(BaseShell):
             carriage_return()
             self._first_prompt = False
         self.settitle()
-        return toks
+        return PygmentsTokens(toks)
 
-    def rprompt_tokens(self, cli):
+    def rprompt_tokens(self):
         """Returns a list of (token, str) tuples for the current right
         prompt.
         """
@@ -216,32 +237,38 @@ class PromptToolkitShell(BaseShell):
         except Exception:  # pylint: disable=broad-except
             print_exception()
         toks = partial_color_tokenize(p)
-        return toks
+        return PygmentsTokens(toks)
 
-    def bottom_toolbar_tokens(self, cli):
+    def _bottom_toolbar_tokens(self):
         """Returns a list of (token, str) tuples for the current bottom
         toolbar.
         """
         p = builtins.__xonsh__.env.get("BOTTOM_TOOLBAR")
-        # self.prompt_formatter does handle empty strings properly,
-        # but this avoids descending into it in the common case of
-        # $TOOLBAR == ''.
-        if isinstance(p, str) and len(p) == 0:
-            return []
+        if not p:
+            return
         try:
             p = self.prompt_formatter(p)
         except Exception:  # pylint: disable=broad-except
             print_exception()
         toks = partial_color_tokenize(p)
-        return toks
+        return PygmentsTokens(toks)
 
-    def continuation_tokens(self, cli, width):
+    @property
+    def bottom_toolbar_tokens(self):
+        """Returns self._bottom_toolbar_tokens if it would yield a result
+        """
+        if builtins.__xonsh__.env.get("BOTTOM_TOOLBAR"):
+            return self._bottom_toolbar_tokens
+
+    def continuation_tokens(self, width, line_number, is_soft_wrap=False):
         """Displays dots in multiline prompt"""
+        if is_soft_wrap:
+            return ""
         width = width - 1
         dots = builtins.__xonsh__.env.get("MULTILINE_PROMPT")
         dots = dots() if callable(dots) else dots
-        if dots is None:
-            return [(Token, " " * (width + 1))]
+        if not dots:
+            return ""
         basetoks = self.format_color(dots)
         baselen = sum(len(t[1]) for t in basetoks)
         if baselen == 0:
@@ -262,7 +289,7 @@ class PromptToolkitShell(BaseShell):
             if n <= count:
                 break
         toks.append((Token, " "))  # final space
-        return toks
+        return PygmentsTokens(toks)
 
     def format_color(self, string, hide=False, force_string=False, **kwargs):
         """Formats a color string using Pygments. This, therefore, returns
@@ -286,17 +313,22 @@ class PromptToolkitShell(BaseShell):
     def print_color(self, string, end="\n", **kwargs):
         """Prints a color string using prompt-toolkit color management."""
         if isinstance(string, str):
-            tokens = partial_color_tokenize(string + end)
+            tokens = partial_color_tokenize(string)
         else:
             # assume this is a list of (Token, str) tuples and just print
             tokens = string
+        tokens = PygmentsTokens(tokens)
         if HAS_PYGMENTS:
             env = builtins.__xonsh__.env
             self.styler.style_name = env.get("XONSH_COLOR_STYLE")
-            proxy_style = PygmentsStyle(pyghooks.xonsh_style_proxy(self.styler))
+            proxy_style = style_from_pygments_cls(
+                pyghooks.xonsh_style_proxy(self.styler)
+            )
         else:
-            proxy_style = style_from_dict(DEFAULT_STYLE_DICT)
-        print_tokens(tokens, style=proxy_style)
+            proxy_style = style_from_pygments_dict(DEFAULT_STYLE_DICT)
+        ptk_print(
+            tokens, style=proxy_style, end=end, include_default_pygments_style=False
+        )
 
     def color_style_names(self):
         """Returns an iterable of all available style names."""
@@ -335,3 +367,9 @@ class PromptToolkitShell(BaseShell):
         #   if not ON_POSIX:
         #       return
         #   sys.stdout.write('\033[9999999C\n')
+        if not ON_POSIX:
+            return
+        stty, _ = builtins.__xonsh__.commands_cache.lazyget("stty", (None, None))
+        if stty is None:
+            return
+        os.system(stty + " sane")
