+++ xontrib/whole_word_jumping.py
@@ -4,8 +4,6 @@ Alt+Left/Right remains unmodified to jump over smaller word segments.
 """
 from prompt_toolkit.keys import Keys
 
-from xonsh.platform import ptk_shell_type
-
 __all__ = ()
 
 
@@ -17,19 +15,14 @@ def custom_keybindings(bindings, **kw):
     # Alt+Left and Alt+Right still jump over smaller word segments.
     # See https://github.com/xonsh/xonsh/issues/2403
 
-    if ptk_shell_type() == "prompt_toolkit2":
-        handler = bindings.add
-    else:
-        handler = bindings.registry.add_binding
-
-    @handler(Keys.ControlLeft)
+    @bindings.add(Keys.ControlLeft)
     def ctrl_left(event):
         buff = event.current_buffer
         pos = buff.document.find_previous_word_beginning(count=event.arg, WORD=True)
         if pos:
             buff.cursor_position += pos
 
-    @handler(Keys.ControlRight)
+    @bindings.add(Keys.ControlRight)
     def ctrl_right(event):
         buff = event.current_buffer
         pos = buff.document.find_next_word_ending(count=event.arg, WORD=True)
